+++ pages/index.js
@@ -8,10 +8,9 @@ const Index = withRouter((props) => {
   return (
     <>
       <Head>
-      <title>CY</title>
-
-        {/* <title>Charmaine Yabsley - Freelance Health Journalist</title> */}
-        </Head>
+        <title>Charmaine Yabsley - Freelance Health Journalist</title>
+        <meta name="Description" content="Charmaine Yabsley - Media Consultant, Writer, Health Journalist"></meta>
+      </Head>
 
       <main className="wide row home">
       <section className="intro-home">