+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+#1.0.2 2019-09-04
+* feautre_audit_tweaks Audit tweaks. html lang to be added
+
+#1.0.1 2019-09-04
+* Slice articles with show all button. Dev mode
+
+#1.0.0 2019-08-05
+* Site launch
+
+#0.0.10 2019-07-26
+* Mobile menu fixes, filter expand collapse
+
 #0.0.10 2019-07-23
 * Store website copy in spreadsheet. Change endpoint to take page number. itemType refactored to itemtype - for React winge